pyproject.toml
[project]
name = "nablax-backend"
version = "0.1.0"
description = "FastAPI email reply service with MCP architecture"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.8.0",
    "openai>=1.3.7",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "mcp[cli]>=1.0.0",
    "openai-agents>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "services"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
]

[tool.uv.workspace]
members = [
    "services/mcp/mail_draft_mcp",
]
